+++ trabalho2/line.cpp
@@ -8,13 +8,13 @@ Line::Line() {
 
 }
 
-Line::Line(GLint& x1, GLint& y1, GLint& x2, GLint& y2, GLfloat* color) {
+Line::Line(GLfloat& x1, GLfloat& y1, GLfloat& x2, GLfloat& y2, GLfloat* color) {
     this->x1 = x1;
     this->y1 = y1;
     this->x2 = x2;
     this->y2 = y2;
 
-    for (GLint i = 0; i < 3; i++) {
+    for (int i = 0; i < 3; i++) {
         this->color[i] = color[i];
     }
 }
